backend/requirements.txt
Django==5.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
google-cloud-bigquery==3.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-filter==23.5
django-cacheops==7.2
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
Pillow==10.1.0 